requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.31
fastapi-sessions==0.3.2
autogen-agentchat~=0.2
pyautogen==0.4.1
pydantic==2.10.3
flaml[automl]
autogen-agentchat[retrievechat-pgvector]~=0.2
# autogen-agentchat[teachable]~=0.2
langchain==0.3.13
#langchain_openai==0.2.1
langchain-community==0.3.11
# chromadb==0.5.21
autogen-agentchat[long-context]~=0.2
azure-cognitiveservices-speech==1.41.1
aioredis==2.0.1 
async-timeout==5.0.1
asyncpg==0.30.0
duckduckgo-search==7.0.2
apify-client==1.8.1
azure-identity==1.13.0b4
azure-ai-formrecognizer==3.2.1
azure-storage-blob==12.14.1
PyPDF2==3.0.1
azure-search-documents==11.5.2
asyncstdlib==3.13.0